[TESTS] utils to change home dir and platform, basic test
utils use pytest.monkeypatch to change Path.home and platform.system.
Added a basic test to check len and content of history returned.

Tests are performed by having a minimal snapshot of the actual contents
of the browser directories. For example, with firefox on linux the
test home directory contains `.mozilla/firefox/profile/places.sqlite` as
the only file. The test home directories will be added in a separate
commit since the `places.sqlite` for firefox can be a few MBs.