[TESTS] test calls CLI function, uses test homedirs
Instead of using subprocesses, CLI function is called directly

This required quite a few changes as the entire thing was written
for subprocesses

Also, added fixtures to perform the tests on all platforms with the
help of test_homedirs. This ensures that the test suite is indepedent
of the tester's system and works even on platforms with no browsers
(such as a CI server). It also makes the tests a LOT faster on systems
where the browser history is huge (such as mine:))