[CLI] Fix profile option issues
This commit handles issue with --show-profiles raising an AttributeError when an unknown browser is given. It also handles a special condition where 'all' is supplied as a parameter to --show-profiles.

It also adds better handling of -b and -p/--profile mutual exclusivity constraints

A typo in get_browser function docs has also been corrected