test: fix test_firefox_linux
Used monkeypatch to make sure that test_firefox_linux still works
despite the new logic added to the `Firefox(Browser)` class for
detecting snap installs. The monkeypatch forces the test to run as if
Firefox was installed using the default (non-snap) method, which matches
the test home-folder hierarchy.

I have not built a test that works with the snap install, since that
installation breaks the paradigm of the profiles being under the home
directory.